requirements.txt
fastapi
uvicorn
mysql-connector-python
numpy
opencv-python
pyzbar
pymupdf
pyjwt
cryptography
python-multipart
sqlalchemy